refactor: update logic imports in frontend
